factor IterableStream into a reusable place
