+++ src/json_stream/requests/__init__.py
@@ -2,26 +2,10 @@ import io
 
 import json_stream
 from json_stream.select_tokenizer import default_tokenizer
+from json_stream.iter_utils import IterableStream
+
 
 CONTENT_CHUNK_SIZE = 10 * 1024
-
-
-class IterableStream(io.RawIOBase):
-    def __init__(self, iterable):
-        self.iterator = iter(iterable)
-        self.remainder = None
-
-    def readinto(self, buffer):
-        try:
-            chunk = self.remainder or next(self.iterator)
-            length = min(len(buffer), len(chunk))
-            buffer[:length], self.remainder = chunk[:length], chunk[length:]
-            return length
-        except StopIteration:
-            return 0    # indicate EOF
-
-    def readable(self):
-        return True
 
 
 def _to_file(response, chunk_size):
