+++ src/json_stream/requests/tests/test_requests.py
@@ -4,21 +4,7 @@ from itertools import zip_longest
 from unittest import TestCase
 from unittest.mock import Mock
 
-from json_stream.requests import IterableStream, load, visit
-
-
-class TestIterableStream(TestCase):
-    def test_read(self):
-        # create some chunks of binary data
-        data = (
-            b"a" * io.DEFAULT_BUFFER_SIZE,
-            b"b" * (io.DEFAULT_BUFFER_SIZE + 1),
-            b"c" * (io.DEFAULT_BUFFER_SIZE - 1),
-        )
-
-        # stream it and check the result
-        stream = IterableStream(data)
-        self.assertEqual(stream.read(), b"".join(data))
+from json_stream.requests import load, visit
 
 
 class TestLoad(TestCase):
